fix: Virtual apps manifest i18n
